refactor: create new `filtergentest` and `clustergentest` packages for test helpers (#824)
Currently some test helpers are located in `clustergen_test` and `filtergen_test` packages. These test helpers need to be imported into other packages in google3, but the current package naming makes it impossible to import them.

Move these test helpers into `clustergentest` and `filtergentest` (no `_`) to allow importing them.
