+++ src/go/configgenerator/filtergen/compressor_test.go
@@ -18,11 +18,12 @@ import (
 	"testing"
 
 	"github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/filtergen"
+	"github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/filtergen/filtergentest"
 	"github.com/GoogleCloudPlatform/esp-v2/src/go/options"
 )
 
 func TestNewCompressorFilterGensFromOPConfig_GenConfig(t *testing.T) {
-	testdata := []SuccessOPTestCase{
+	testdata := []filtergentest.SuccessOPTestCase{
 		{
 			Desc: "Generate with compression enabled",
 			OptsIn: options.ConfigGeneratorOptions{
