+++ src/go/configgenerator/filtergen/health_check_test.go
@@ -18,12 +18,13 @@ import (
 	"testing"
 
 	"github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/filtergen"
+	"github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/filtergen/filtergentest"
 	"github.com/GoogleCloudPlatform/esp-v2/src/go/options"
 	confpb "google.golang.org/genproto/googleapis/api/serviceconfig"
 )
 
 func TestNewHealthCheckFilterGensFromOPConfig_GenConfig(t *testing.T) {
-	testdata := []SuccessOPTestCase{
+	testdata := []filtergentest.SuccessOPTestCase{
 		{
 			Desc: "Success, generate health check filter for standard /healthz path",
 			OptsIn: options.ConfigGeneratorOptions{
